learn-agent-in-24hours/requirements.txt
openai
python-dotenv
pydantic
langchain
langchain-openai
langgraph
chromadb
fastapi[standard]
streamlit
mcp
